리뷰 4개 : 컨트롤러 로퍼지 서비스 서임플 & 내꺼.

+++ FIB/src/main/java/com/fox/fib/controller/RestProductController.java
@@ -149,59 +149,67 @@ public class RestProductController {
 	public ResponseEntity<?> recentProduct(@RequestParam(name = "id") String id, @RequestParam(name = "pcode") String pcode,
 		Product pentity, RecentView rentity) {
 		try {
-			log.info("[153]최근방문 pcode 확인 : " + pcode);
+			log.info("[153]최근방문 상품 확인 : " + id + " & " + pcode);
 
-			int product_code = Integer.parseInt(pcode);
+			if (id != null && !id.isEmpty()) {
 
-			pentity = productservice.selectOne(product_code);
+				int product_code = Integer.parseInt(pcode);
 
-			if (recentviewservice.checkDuplicated(id, product_code) == 0) {
+				pentity = productservice.selectOne(product_code);
 
-				List<RecentView> recentviewList = recentviewservice.selectListForUserId(id);
+				if (recentviewservice.checkDuplicated(id, product_code) == 0) {
 
-				if (recentviewList.size() >= 5) {
-					recentviewservice.deleteOldest(id);
-				}
+					List<RecentView> recentviewList = recentviewservice.selectListForUserId(id);
 
-				rentity.setId(id);
-				rentity.setProduct_code(product_code);
-				rentity.setProtype(pentity.getProtype());
-				rentity.setTitle(pentity.getTitle());
-				rentity.setImage(pentity.getImage());
-				rentity.setPrice(pentity.getPrice());
+					if (recentviewList.size() >= 5) {
+						recentviewservice.deleteOldest(id);
+					}
 
-				recentviewservice.save(rentity);
+					rentity.setId(id);
+					rentity.setProduct_code(product_code);
+					rentity.setProtype(pentity.getProtype());
+					rentity.setTitle(pentity.getTitle());
+					rentity.setImage(pentity.getImage());
+					rentity.setPrice(pentity.getPrice());
 
-				log.info("새 최근본 상품 세이브!");
+					recentviewservice.save(rentity);
 
+					log.info("새 최근본 상품 세이브!");
 
-			} else if (recentviewservice.checkDuplicated(id, product_code) == 1) {
+				} else if (recentviewservice.checkDuplicated(id, product_code) == 1) {
 
-				log.info("중복 제품이라 안 들어감");
-				return ResponseEntity.ok("중복 제품이라 안 들어감");
+					log.info("중복 제품이라 안 들어감");
+					return ResponseEntity.ok("중복 제품이라 안 들어감");
+
+				} // else if
+
+			} else {
+				return ResponseEntity.ok("로그인 후 이용 가능");
 			}
+
 		} catch (Exception e) {
-			log.info("최근방문상품 저장 실패요 : " + e.toString());
+			log.error("최근방문상품 저장 실패요 : " + e.toString());
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("최근방문 상품데이터 저장 실패");
 		}
-
-		return ResponseEntity.ok("최근본 상품 잘 들어감");
-
+		return ResponseEntity.ok("예외없음");
 	}
 
 	// ======================================================================================================
 
 	@GetMapping("/selectRecentProductList")
-	public List<RecentView> selectRecentProductList(@RequestParam(name = "id") String id, RecentView rentity) {
+	public ResponseEntity<?> selectRecentProductList(@RequestParam(name = "id") String id, RecentView rentity) {
 
 		try {
-			List<RecentView> recentviewList = recentviewservice.selectListForUserId(id);
-
-			return recentviewList;
+			if (id == null || id.isEmpty()) {
+				return ResponseEntity.ok("로그인 후 확인하실 수 있습니다.");
+			} else {
+				List<RecentView> recentviewList = recentviewservice.selectListForUserId(id);
+				return ResponseEntity.ok(recentviewList);
+			}
 
 		} catch (Exception e) {
 			log.info("최근방문상품 실패요 : " + e.toString());
-			return null;
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("최근방문상품출력 예외오류");
 		}
 
 	}
